Add tests and fix default constructor
Fix default constructor case of Hour(24) when applied to a Date vector.
Hour cannot be added to Date objects, so the step must be converted to
Day, but because this is a change to the step hyper parameter it must be
adjusted. Unfortunately, this case cannot be distinguished from a user
supplied date, so a warning must be shown, which is non-ideal for a
common default case.

Alternatively we can add an extra internal variable to track whether a
user supplied step is given and only warn on that case, but that is
extra complexity I am avoiding for this commit.